# Optimize parsers aggregating multiple values in a list