Using `Value.aggregate` to append an element to a list implicitly creates a new copy of the list every time. When parsing a large enough number of values (e.g. 100000), the parsing time ends up dominated by the time it takes to copy the list.

This PR instead uses a single mutable list and appends to it.

### Benchmark examples

1.
```
p = sepBy(string('a'), string(','))
s = ','.join('a'*100000)
```
`p.parse(s)` took 59 seconds on my machine before, only 3.6 seconds with the fix.

2.
```
p = many(string('a'))
s = 'a'*100000
```
`p.parse(s)` took 57 seconds on my machine before, only 2 seconds with the fix.